+++ admin/partials/decalog-admin-settings-logger-edit.php
@@ -23,7 +23,7 @@
 				'handler' => $current_logger['handler'],
 				'uuid'    => $current_logger['uuid'],
 			),
-			admin_url( 'options-general.php' )
+			admin_url( 'admin.php' )
 		)
 	);
 	?>