+++ includes/plugin/class-core.php
@@ -86,6 +86,7 @@ class Core {
 		$listeners = new ListenerFactory();
 		$this->loader->add_filter( 'perfopsone_plugin_info', self::class, 'perfopsone_plugin_info' );
 		$this->loader->add_action( 'init', $bootstrap, 'initialize', 0 );
+		$this->loader->add_action( 'init', $bootstrap, 'late_initialize', PHP_INT_MAX );
 		$this->loader->add_action( 'plugins_loaded', $listeners, 'launch', 1 );
 		$this->loader->add_action( 'plugins_loaded', $listeners, 'launch_late_init', PHP_INT_MAX );
 		$this->loader->add_action( 'wp_head', $assets, 'prefetch' );