+++ includes/system/class-useragent.php
@@ -56,7 +56,7 @@ class UserAgent {
 		foreach ( $fields as $key => $arg ) {
 			$params[ $key ] = $arg;
 		}
-		$url = admin_url( 'tools.php?page=podd-viewer' );
+		$url = admin_url( 'admin.php?page=podd-viewer' );
 		foreach ( $params as $key => $arg ) {
 			if ( '' !== $arg ) {
 				$url .= '&' . $key . '=' . rawurlencode( $arg );