+++ perfopsone/autoload.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Autoload for Sessions.
+ *
+ * @package Bootstrap
+ * @author  Pierre Lannoy <https://pierre.lannoy.fr/>.
+ * @since   1.0.0
+ */
+
+spl_autoload_register(
+	function ( $class ) {
+		$classname = $class;
+		$filepath  = __DIR__ . '/';
+		if ( strpos( $classname, 'PerfOpsOne\\' ) === 0 ) {
+			while ( strpos( $classname, '\\' ) !== false ) {
+				$classname = substr( $classname, strpos( $classname, '\\' ) + 1, 1000 );
+			}
+			$filename = 'class-' . str_replace( '_', '-', strtolower( $classname ) ) . '.php';
+			$file     = $filepath . $filename;
+			if ( file_exists( $file ) ) {
+				include_once $file;
+			}
+		}
+	}
+);